Significant refactor, breaks backward compat
