+++ psycho_embeddings/__init__.py
@@ -1,5 +1,7 @@
 """Top-level package for Psycho Embeddings."""
 
 __author__ = """Federico Bianchi"""
-__email__ = 'f.bianchi@unibocconi.it'
-__version__ = '0.1.0'
+__email__ = "f.bianchi@unibocconi.it"
+__version__ = "0.1.0"
+
+from .embedding import BERTEmbedder
